fix(BaseComponent): :truck: BaseButton2 >> BaseButton으로 이름 변경
